Refactoring User View code
